Reworked the utility class method
