+++ test/Chirp.Web.Tests/E2ETests.cs
@@ -1,6 +1,67 @@
-﻿namespace Chirp.Web.Tests;
+﻿using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.VisualStudio.TestPlatform.TestHost;
+using Xunit;
 
-public class E2ETests
+namespace Chirp.Web.Tests;
+
+public class E2ETests  : IClassFixture<WebApplicationFactory<Program>>
 {
-    
+    /*
+  private readonly WebApplicationFactory<Program> _fixture;
+  private readonly HttpClient _client;
+
+  public IntegrationTests(WebApplicationFactory<Program> fixture)
+  {
+      
+      _fixture = fixture;
+      _client = _fixture.CreateClient(new WebApplicationFactoryClientOptions { AllowAutoRedirect = true, HandleCookies = true });
+  }
+
+  [Fact]
+  public async void CanSeePublicTimeline()
+  {
+      var response = await _client.GetAsync("/");
+      response.EnsureSuccessStatusCode();
+      var content = await response.Content.ReadAsStringAsync();
+
+      Assert.Contains("Chirp!", content);
+      Assert.Contains("Public Timeline", content);
+  }
+
+  [Theory]
+  [InlineData("Helge")]
+  [InlineData("Adrian")]
+  public async void CanSeePrivateTimeline(string author)
+  {
+      var response = await _client.GetAsync($"/{author}");
+      response.EnsureSuccessStatusCode();
+      var content = await response.Content.ReadAsStringAsync();
+
+      Assert.Contains("Chirp!", content);
+      Assert.Contains($"{author}'s Timeline", content);
+  }
+
+  [Theory]
+  [InlineData("Helge", "Hello, BDSA students!")]
+  [InlineData("Adrian","Hej, velkommen til kurset." )]
+  public async void PrivateTimelineShowsContent(string author, string message)
+  {
+      var response = await _client.GetAsync($"/{author}");
+      response.EnsureSuccessStatusCode();
+      var content = await response.Content.ReadAsStringAsync();
+
+      Assert.Contains(message, content);
+  }
+
+  [Fact]
+  public async void PublicTimelineShowsContent()
+  {
+      var response = await _client.GetAsync("/");
+      response.EnsureSuccessStatusCode();
+      var content = await response.Content.ReadAsStringAsync();
+
+      Assert.Contains("Starbuck now is what we hear the worst", content);
+  }
+
+  */
 }
