+++ test/Chirp.Web.Tests/IntegrationTests.cs
@@ -5,64 +5,20 @@ using Xunit;
 namespace Chirp.Web.Tests;
 
 
-public class IntegrationTests : IClassFixture<WebApplicationFactory<Program>>
+public class IntegrationTests
 {   
     /*
-    private readonly WebApplicationFactory<Program> _fixture;
-    private readonly HttpClient _client;
-
-    public IntegrationTests(WebApplicationFactory<Program> fixture)
-    {
-        //TODO: ensure we use testDB for this (after we get dependency injection working)
-        _fixture = fixture;
-        _client = _fixture.CreateClient(new WebApplicationFactoryClientOptions { AllowAutoRedirect = true, HandleCookies = true });
-    }
-
     [Fact]
-    public async void CanSeePublicTimeline()
+    public async Task TestAddCheep()
     {
-        var response = await _client.GetAsync("/");
-        response.EnsureSuccessStatusCode();
-        var content = await response.Content.ReadAsStringAsync();
+        var repository = await TestUtilities.createInMemoryDB();
 
-        Assert.Contains("Chirp!", content);
-        Assert.Contains("Public Timeline", content);
-    }
-
-    [Theory]
-    [InlineData("Helge")]
-    [InlineData("Adrian")]
-    public async void CanSeePrivateTimeline(string author)
-    {
-        var response = await _client.GetAsync($"/{author}");
-        response.EnsureSuccessStatusCode();
-        var content = await response.Content.ReadAsStringAsync();
-
-        Assert.Contains("Chirp!", content);
-        Assert.Contains($"{author}'s Timeline", content);
-    }
-    
-    [Theory]
-    [InlineData("Helge", "Hello, BDSA students!")]
-    [InlineData("Adrian","Hej, velkommen til kurset." )]
-    public async void PrivateTimelineShowsContent(string author, string message)
-    {
-        var response = await _client.GetAsync($"/{author}");
-        response.EnsureSuccessStatusCode();
-        var content = await response.Content.ReadAsStringAsync();
+        var cheepsBefore = await repository.GetCheepsFromAuthor(0, "Mellie Yost");
         
-        Assert.Contains(message, content);
-    }
-
-    [Fact]
-    public async void PublicTimelineShowsContent()
-    {
-        var response = await _client.GetAsync("/");
-        response.EnsureSuccessStatusCode();
-        var content = await response.Content.ReadAsStringAsync();
+        await repository.AddCheep("hiii" );
         
-        Assert.Contains("Starbuck now is what we hear the worst", content);
-    }
-    
-    */
+        var cheepsAfter = await repository.GetCheepsFromAuthor(0, "Mellie Yost");
+        
+        Assert.True(cheepsBefore.Count != cheepsAfter.Count);
+    } */
 }
